feat: добавить пакет storage, интерфейсы и классы
