+++ pom.xml
@@ -45,6 +45,12 @@
 			<artifactId>logbook-spring-boot-starter</artifactId>
 			<version>2.14.0</version>
 		</dependency>
+		<dependency>
+			<groupId>org.jetbrains</groupId>
+			<artifactId>annotations</artifactId>
+			<version>20.1.0</version>
+			<scope>compile</scope>
+		</dependency>
 	</dependencies>
 
 	<build>
