+++ src/main/java/ru/yandex/practicum/filmorate/exception/ValidationException.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.controller;
+package ru.yandex.practicum.filmorate.exception;
 
 public class ValidationException extends RuntimeException{
     public ValidationException(String message) {
