+++ src/main/java/ru/yandex/practicum/filmorate/storage/FilmStorage.java
@@ -1,0 +1,13 @@
+package ru.yandex.practicum.filmorate.storage;
+
+import ru.yandex.practicum.filmorate.model.Film;
+
+import java.util.Collection;
+
+public interface FilmStorage {
+    Film add(Film film);
+
+    Film update(Film film);
+
+    Collection<Film> getAll();
+}
