+++ src/main/java/ru/yandex/practicum/filmorate/storage/InMemoryUserStorage.java
@@ -1,0 +1,69 @@
+package ru.yandex.practicum.filmorate.storage;
+
+import lombok.Getter;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Component;
+import ru.yandex.practicum.filmorate.exception.ValidationException;
+import ru.yandex.practicum.filmorate.model.User;
+
+import java.time.LocalDate;
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.Map;
+
+@Slf4j
+@Component
+public class InMemoryUserStorage implements UserStorage{
+
+    private int userId = 1;
+
+    @Getter
+    private final Map<Integer, User> users = new HashMap<>();
+
+    @Override
+    public User add(User user) {
+        check(user);
+        user.setId(userId++);
+        users.put(user.getId(), user);
+        return user;
+    }
+
+    @Override
+    public User update(User user) {
+        if (!users.containsKey(user.getId())) {
+            log.error("Пользователь {} не найден", user);
+            throw new ValidationException("Пользователь с таким id не найден.");
+        }
+        check(user);
+        users.put(user.getId(), user);
+        return user;
+    }
+
+    @Override
+    public Collection<User> getAll() {
+        return users.values();
+    }
+
+    public void check(User user) {
+        if (user.getEmail() == null || user.getEmail().isBlank()) {
+            log.error("Адрес электронной почты не может быть пустым.");
+            throw new ValidationException("Адрес электронной почты не может быть пустым.");
+        }
+        if (!user.getEmail().contains("@")) {
+            log.error("Адрес электронной почты указан в неверном формате.");
+            throw new ValidationException("Адрес электронной почты указан в неверном формате.");
+        }
+        if (user.getLogin() == null || user.getLogin().isBlank() || user.getLogin().contains(" ")) {
+            log.error("Логин не должен быть пустым или содержать пробелы");
+            throw new ValidationException("Логин не должен быть пустым или содержать пробелы");
+        }
+        if (user.getBirthday().isAfter(LocalDate.now())) {
+            log.error("Дата рождения не может быть указана позже текущей даты.");
+            throw new ValidationException("Дата рождения не может быть указана позже текущей даты.");
+        }
+        if (user.getName() == null || user.getName().isBlank()) {
+            user.setName(user.getLogin());
+        }
+
+    }
+}
